Switched to official TinyMCE React integration.
