+++ .changeset/silver-socks-dress.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/fields-wysiwyg-tinymce': minor
+---
+
+Switched to official TinyMCE React integration.
