+++ packages/fields-wysiwyg-tinymce/package.json
@@ -18,6 +18,7 @@
     "@emotion/core": "^10.0.27",
     "@keystonejs/build-field-types": "^5.2.1",
     "@keystonejs/fields": "^6.3.1",
+    "@tinymce/tinymce-react": "^3.5.0",
     "cuid": "^2.1.6",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
