+++ packages/fields-wysiwyg-tinymce/src/views/Field.js
@@ -1,35 +1,70 @@
 /** @jsx jsx */
 
-import { jsx } from '@emotion/core';
-import { Component } from 'react';
+import { Global, css, jsx } from '@emotion/core';
 
 import { FieldContainer, FieldLabel } from '@arch-ui/fields';
 
-import Editor from './Editor';
+// MUST IMPORT for TinyMCE to work!
+// eslint-disable-next-line no-unused-vars
+import tinymce from 'tinymce/tinymce';
 
-export default class WysiwygField extends Component {
-  onChange = value => {
-    if (typeof value === 'string' && value !== this.props.value) {
-      this.props.onChange(value);
+import { Editor } from '@tinymce/tinymce-react';
+
+const defaultOptions = {
+  autoresize_bottom_margin: 20,
+  base_url: '/tinymce-assets',
+  branding: false,
+  menubar: false,
+  plugins: 'link lists code autoresize paste quickbars hr table emoticons image',
+  statusbar: false,
+  toolbar:
+    'formatselect forecolor | alignleft aligncenter alignright alignjustify | bullist numlist indent outdent | link unlink | image table emoticons hr | code',
+  quickbars_selection_toolbar:
+    'bold italic underline strikethrough | h1 h2 h3 | quicklink blockquote removeformat',
+  quickbars_insert_toolbar: false,
+  width: '100%',
+};
+
+// Overrides some of the TinyMCE theme to better match the Admin UI style
+const GlobalStyles = () => (
+  <Global
+    styles={css`
+      .tox-tinymce {
+        border-radius: 5px !important;
+        border-color: #c1c7d0 !important;
+      }
+    `}
+  />
+);
+
+const WysiwygField = ({ onChange, autoFocus, field, errors, value: serverValue }) => {
+  const handleChange = value => {
+    if (typeof value === 'string') {
+      onChange(value);
     }
   };
-  render() {
-    const { autoFocus, field, errors, value: serverValue } = this.props;
-    const value = serverValue || '';
-    const htmlID = `ks-input-${field.path}`;
-    const accessError = errors.find(
-      error => error instanceof Error && error.name === 'AccessDeniedError'
-    );
 
-    if (accessError) return null;
+  const value = serverValue || '';
+  const htmlID = `ks-input-${field.path}`;
+  const accessError = errors.find(
+    error => error instanceof Error && error.name === 'AccessDeniedError'
+  );
 
-    return (
-      <FieldContainer>
-        <FieldLabel htmlFor={htmlID} field={field} errors={errors} />
-        <div css={{ display: 'flex', flex: 1 }}>
-          <Editor value={value} onChange={this.onChange} id={htmlID} autoFocus={autoFocus} />
-        </div>
-      </FieldContainer>
-    );
-  }
-}
+  if (accessError) return null;
+
+  return (
+    <FieldContainer>
+      <FieldLabel htmlFor={htmlID} field={field} errors={errors} />
+      <div css={{ display: 'flex', flex: 1 }}>
+        <GlobalStyles />
+        <Editor
+          init={{ ...defaultOptions, auto_focus: autoFocus }}
+          onEditorChange={handleChange}
+          value={value}
+        />
+      </div>
+    </FieldContainer>
+  );
+};
+
+export default WysiwygField;
